Final Commit
Ready for Review. Changed the wc_hole01 to index.html. Renamed the wc_styles.css. Added in styles to display the body as a flex. Added styles to display the article as a flex. Added styles to display the aside as a flex. Added styles to display the navigation bar as a flex for desktop/tablet. Added styles to hide the navigation bar on mobile devices. Changed all of the links on all of the pages that had wc_hole01 to index.